+++ demo/node/src/chain_spec.rs
@@ -1,11 +1,12 @@
+use authority_selection_inherents::CommitteeMember;
+use partner_chains_cli::{CreateChainSpecConfig, ParsedPermissionedCandidatesKeys};
 use partner_chains_demo_runtime::{
 	AccountId, CrossChainPublic, Signature, WASM_BINARY, opaque::SessionKeys,
 };
+use sc_service::ChainType;
+use sidechain_slots::SlotsPerEpoch;
 use sp_core::{Pair, Public};
 use sp_runtime::traits::{IdentifyAccount, Verify};
-
-// The URL for the telemetry server.
-// const STAGING_TELEMETRY_URL: &str = "wss://telemetry.polkadot.io/submit/";
 
 pub type ChainSpec = sc_service::GenericChainSpec;
 
@@ -33,4 +34,56 @@ where
 
 pub fn runtime_wasm() -> &'static [u8] {
 	WASM_BINARY.expect("Runtime wasm not available")
+}
+
+fn permissioned_candidate_to_committee_member(
+	keys: &ParsedPermissionedCandidatesKeys,
+) -> CommitteeMember<CrossChainPublic, SessionKeys> {
+	let session_keys = (keys.aura, keys.grandpa).into();
+	CommitteeMember::permissioned(keys.sidechain.try_into().unwrap(), session_keys)
+}
+
+fn permissioned_candidate_to_pallet_partner_chains_session_keys(
+	keys: &ParsedPermissionedCandidatesKeys,
+) -> (AccountId, SessionKeys) {
+	let session_keys = (keys.aura, keys.grandpa).into();
+	(keys.account_id_32(), session_keys)
+}
+
+/// Creates chain-spec according to the config obtained by wizards.
+/// [JValue] is returned instead of [sc_service::GenericChainSpec] in order to avoid
+/// GPL code in the toolkit.
+pub fn pc_create_chain_spec(config: &CreateChainSpecConfig) -> serde_json::Value {
+	let runtime_genesis_config = partner_chains_demo_runtime::RuntimeGenesisConfig {
+		system: partner_chains_demo_runtime::SystemConfig::default(),
+		balances: partner_chains_demo_runtime::BalancesConfig::default(),
+		aura: partner_chains_demo_runtime::AuraConfig::default(),
+		grandpa: partner_chains_demo_runtime::GrandpaConfig::default(),
+		sudo: partner_chains_demo_runtime::SudoConfig::default(),
+		transaction_payment: Default::default(),
+		session: config.pallet_partner_chains_session_config(
+			permissioned_candidate_to_pallet_partner_chains_session_keys,
+		),
+		sidechain: config.pallet_sidechain_config(SlotsPerEpoch::default()),
+		pallet_session: Default::default(),
+		session_committee_management: config
+			.pallet_session_validator_management_config(permissioned_candidate_to_committee_member),
+		native_token_management: config.native_token_management_config(),
+		governed_map: config.governed_map_config(),
+		test_helper_pallet: partner_chains_demo_runtime::TestHelperPalletConfig {
+			participation_data_release_period: 30,
+			..Default::default()
+		},
+	};
+	let genesis_json = serde_json::to_value(runtime_genesis_config)
+		.expect("Genesis config must be serialized correctly");
+	let chain_spec = ChainSpec::builder(runtime_wasm(), None)
+		.with_name("Partner Chains Demo")
+		.with_id("partner_chains_demo")
+		.with_chain_type(ChainType::Live)
+		.with_genesis_config(genesis_json)
+		.build();
+	let raw = false;
+	let chain_spec_str = chain_spec.as_json(raw).expect("Chain spec serialization can not fail");
+	serde_json::from_str(&chain_spec_str).unwrap()
 }